Rest_api_service/requirements.txt
aniso8601==9.0.1
annotated-types==0.5.0
attrs==23.1.0
autocommand==2.2.2
biopython==1.81
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
cheroot==10.0.0
CherryPy==18.8.0
click==7.1.2
Flask==1.1.4
flask-restplus==0.13.0
future==0.18.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflect==7.0.0
itsdangerous==1.1.0
jaraco.collections==4.3.0
jaraco.context==4.3.0
jaraco.functools==3.8.0
jaraco.text==3.11.1
Jinja2==2.11.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.0.1
more-itertools==10.0.0
numpy==1.24.4
pandas==2.0.3
Paste==3.5.3
pkgutil_resolve_name==1.3.10
portend==3.2.0
presto==0.7.1
presto-python-client==0.8.3
pydantic==2.1.1
pydantic_core==2.4.0
PyHive==0.6.5
python-dateutil==2.8.2
pytz==2023.3
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
scipy==1.10.1
six==1.16.0
tempora==5.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==0.16.1
zc.lockfile==3.0.post1
zipp==3.16.2
